[WebGPU] Make Parser class declaration visible to unit tests for testing https://bugs.webkit.org/show_bug.cgi?id=241738
Patch by Kiet Ho <tho22@apple.com> on 2022-06-21
Reviewed by Myles C. Maxfield.

The existing parser interface (WGSL::parseLChar() and WGSL::parseUChar())
only accepts entirely valid shader module, which makes it difficult to
write unit tests targeting individual aspects of the parser (for example,
only testing the type declaration or unary expression parsing logic). This
commit moves the Parser class declaration from the .cpp file to a private
header file (ParserPrivate.h), and makes private methods that parse
individual grammar rules public. This gives unit tests access to those
methods to test them.

* Source/WebGPU/WGSL/Parser.cpp:
(WGSL::Parser::Parser): Deleted.
(WGSL::Parser::current): Deleted.
* Source/WebGPU/WGSL/ParserPrivate.h: Added.
(WGSL::Parser::Parser):
(WGSL::Parser::current):
* Source/WebGPU/WebGPU.xcodeproj/project.pbxproj:

Canonical link: https://commits.webkit.org/251717@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@295712 268f45cc-cd09-0410-ab3c-d52691b4dbfc
